requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.4
django-filter==22.1
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
flake8==6.0.0
idna==3.4
kombu==5.2.4
Markdown==3.4.1
mccabe==0.7.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==0.21.0
pytz==2022.7
PyYAML==6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
ujson==5.7.0
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6

pytest~=7.2.2